+++ CHANGELOG.md
@@ -9,10 +9,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - n/a
 
+## [0.2.0] · 2020-12-06
+
+### Added
+- Throws an exception if `expect` was called but no assertion was made.
+
+
 ## [0.1.1] · 2020-12-05
 
 ### Fixed
-- Fixed `toBe` so that it's legal iff `received is expected` is legal.
+- `toBe` is now legal iff `received `**`is`**` expected` is legal.
 
 
 ## [0.1.0] · 2020-11-22
@@ -26,3 +32,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 [0.1.0]: https://gitlab.com/andrej88/exceeds-expectations/-/tree/v0.1.0
 [0.1.1]: https://gitlab.com/andrej88/exceeds-expectations/-/tree/v0.1.1
+[0.2.0]: https://gitlab.com/andrej88/exceeds-expectations/-/tree/v0.2.0